Adding logging/tracing to Azure Distributed file system
Logging/Tracing must be injected into ServiceProvider and later be used by components.
TracingService will be using objects which implement ITraceEvent. ITraceEvent can either be a root event or chain of events. If the chain is created properly, all the code paths which execution has happened will be traced and logged correctly.
TracingService will be using the same logging service which is provided in this JIRA to log the traces.
LoggingService will be responsible to get all the logs from AzureDistributedFileSystem components and dump them onto Hadoop Logger in different levels.
Unit tests added.