Adding Azure Distributed file system auto-generated code and implementing related services
Auto generated com.microsoft.azure.dfs.client package using auto-rest and modified the code to fix stylechecks errors.
Added a couple of new service to be able to attach authorization header to the request and send the request to the service.
The rest client service will be extracted from the AzureDistributedFileSystemImpl in the next code review.
Added basic test coverage.